backend/pyproject.toml
[tool.poetry]
name = "ecommerce-backend"
version = "0.1.0"
description = "E-Commerce + Dashboard (AR / Mercado Pago) — Backend"
authors = ["Antonio Romero <antonio@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.120.4"
uvicorn = "0.30.1"
starlette = "0.49.1"
pydantic = "2.12.1"
pydantic-settings = "2.6.1"
SQLAlchemy = "2.0.43"
alembic = "1.14.0"
PyMySQL = "1.1.1"
redis = "5.0.8"
celery = "5.4.0"
httpx = "0.28.1"
python-multipart = "0.0.17"
jinja2 = "3.1.4"
email-validator = "2.2.0"
python-jose = "3.3.0"
passlib = "1.7.4"
mercadopago = "2.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
ruff = "0.6.9"
mypy = "1.13.0"

